Fixed selectivityEstimate type bug

+++ arango_driver/lib/src/arango_db_client.dart
@@ -847,7 +847,7 @@ class ArangoDBClient {
         unique: map['unique'],
         deduplicate: map['deduplicate'],
         sparse: map['sparse'],
-        selectivityEstimate: map['selectivityEstimate'],
+        selectivityEstimate: map['selectivityEstimate'].toDouble(),
       ),
     );
   }
